LayoutTests/storage/indexeddb/modern/deleteobjectstore-1-expected.txt

This test creates an object store then populates it, then commits that transaction. It then deletes it, but aborts that transaction. Finally it checks to make sure everything from step 1 is there as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


ALERT: Initial upgrade needed: Old version - 0 New version - 1
ALERT: Initial upgrade versionchange transaction complete
ALERT: Second upgrade needed: Old version - 1 New version - 2
ALERT: Deleting object store
ALERT: Second version change transaction abort
ALERT: Request error - [object Event]
ALERT: Success opening database connection - Starting final transaction
ALERT: Value gotten was AH AH AH AH AH
ALERT: Value gotten was AH AH AH AH AH
ALERT: Value gotten was AH AH AH AH AH
ALERT: Value gotten was AH AH AH AH AH
ALERT: Value gotten was AH AH AH AH AH
ALERT: Value gotten was AH AH AH AH AH
ALERT: Value gotten was AH AH AH AH AH
ALERT: Value gotten was AH AH AH AH AH
ALERT: Value gotten was AH AH AH AH AH
ALERT: Value gotten was AH AH AH AH AH
ALERT: Final transaction complete
PASS successfullyParsed is true

TEST COMPLETE
